chore: Removed `chart.js` and legacy components from front end
