+++ front-end-components/package.json
@@ -16,13 +16,11 @@
     "lint:fix": "eslint . --ext ts,tsx,cjs --fix"
   },
   "dependencies": {
-    "chart.js": "^4.4.2",
     "classnames": "^2.5.1",
     "file-saver": "^2.0.5",
     "govuk-frontend": "^5.2.0",
     "punycode": "^2.3.1",
     "react": "^18.2.0",
-    "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.2.0",
     "recharts": "^2.12.1",
     "recharts-to-png": "^2.3.1",
