--- front-end-components/src/components/charts/horizontal-bar-chart-2/index.tsx
@@ -1,3 +1,0 @@
-/* eslint-disable react-refresh/only-export-components */
-export * from "src/components/charts/horizontal-bar-chart-2/types";
-export * from "src/components/charts/horizontal-bar-chart-2/component";
