+++ front-end-components/src/components/charts/horizontal-bar-chart/component.tsx
@@ -1,153 +1,227 @@
-import { forwardRef, useContext, useImperativeHandle, useRef } from "react";
-import { Bar } from "react-chartjs-2";
 import {
-  Chart as ChartJS,
-  CategoryScale,
-  LinearScale,
-  BarElement,
-  Title,
-  Tick,
-  ChartOptions,
-} from "chart.js";
-import { ChartDimensionContext, SelectedSchoolContext } from "src/contexts";
+  ForwardedRef,
+  ReactElement,
+  Ref,
+  forwardRef,
+  useImperativeHandle,
+  useMemo,
+  useState,
+} from "react";
 import {
-  BarChartProps,
-  DownloadHandle,
+  HorizontalBarChartProps,
+  LabelListContentProps,
 } from "src/components/charts/horizontal-bar-chart";
+import {
+  Bar,
+  BarChart,
+  CartesianGrid,
+  Cell,
+  Label,
+  LabelList,
+  Legend,
+  ResponsiveContainer,
+  Text,
+  Tooltip,
+  XAxis,
+  YAxis,
+} from "recharts";
+import { CategoricalChartState } from "recharts/types/chart/types";
+import classNames from "classnames";
+import {
+  ChartDataSeries,
+  ChartHandler,
+  ChartSeriesValue,
+} from "src/components";
+import { useDownloadPngImage } from "src/hooks/useDownloadImage";
 
-ChartJS.register(CategoryScale, LinearScale, BarElement, Title);
+function HorizontalBarChartInner<TData extends ChartDataSeries>(
+  props: HorizontalBarChartProps<TData>,
+  ref: ForwardedRef<ChartHandler>
+) {
+  const {
+    barCategoryGap,
+    chartName,
+    data,
+    grid,
+    hideXAxis,
+    hideYAxis,
+    highlightActive,
+    highlightedItemKeys,
+    keyField,
+    labels,
+    legend,
+    margin: _margin,
+    onImageLoading,
+    seriesConfig,
+    seriesLabelField,
+    tick,
+    tickWidth,
+    tooltip,
+    valueFormatter,
+    valueLabel,
+    valueUnit,
+  } = props;
 
-const underLinePlugin = {
-  id: "underline",
-  afterDraw: (chart: ChartJS) => {
-    const ctx = chart.ctx;
-    ctx.save();
+  const { downloadPng, ref: rechartsRef } = useDownloadPngImage({
+    fileName: `${chartName}.png`,
+    onImageLoading,
+  });
 
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const yAxis = chart.scales.y as any;
+  useImperativeHandle(ref, () => ({
+    download() {
+      downloadPng();
+    },
+  }));
 
-    yAxis.ticks.forEach((tick: Tick, index: number) => {
-      const yAxisLabelPadding = yAxis.options.ticks.padding;
-      const textWidth = ctx.measureText(tick.label as string).width;
-      const yPosition = yAxis.getPixelForTick(index) + yAxisLabelPadding * 2;
+  const visibleSeriesNames: (keyof TData)[] = useMemo(() => {
+    return seriesConfig
+      ? Object.entries(seriesConfig)
+          .filter((v) => v[1]?.visible)
+          .map((v) => v[0])
+      : Object.keys(data).filter(
+          (k) => k !== seriesLabelField && k !== keyField
+        );
+  }, [data, keyField, seriesConfig, seriesLabelField]);
 
-      ctx.strokeStyle = yAxis.options.ticks.color;
-      ctx.lineWidth = 1.5;
-      const xStart = yAxis.right - textWidth - yAxisLabelPadding * 3;
+  const [activeItemIndex, setActiveItemIndex] = useState<number>();
+  const handleBarChartMouseMove = (nextState: CategoricalChartState) => {
+    setActiveItemIndex(nextState.activeTooltipIndex);
+  };
 
-      ctx.beginPath();
-      ctx.moveTo(xStart, yPosition);
-      ctx.lineTo(xStart + textWidth, yPosition);
-      ctx.stroke();
-    });
-    ctx.restore();
-  },
-};
+  const margin = _margin || 5;
 
-export const HorizontalBarChart = forwardRef<DownloadHandle, BarChartProps>(
-  (props, ref) => {
-    const { data, chartName } = props;
-    const labels = data.map((dataPoint) => dataPoint.school);
-    const values = data.map((dataPoint) => dataPoint.value);
-    const selectedSchool = useContext(SelectedSchoolContext);
-    const xLabel = useContext(ChartDimensionContext);
-    const chartContainerStyle = {
-      width: "100%",
-      minHeight: `${labels.length * 30}px`,
-    };
+  // https://stackoverflow.com/a/61373602/504477
+  const renderCell = (
+    entry: TData,
+    dataIndex: number,
+    seriesKey: keyof TData,
+    seriesIndex: number
+  ) => {
+    const config = seriesConfig && seriesConfig[seriesKey];
 
-    const chosenSchoolIndex = selectedSchool
-      ? labels.indexOf(selectedSchool.name)
-      : 0;
-    const barBackgroundColors = labels.map((_, index) =>
-      index === chosenSchoolIndex ? "#12436D" : "#BFBFBF"
-    );
-
-    const chartRef = useRef<ChartJS<"bar">>(null);
-
-    const datasets = [
+    const className = classNames(
+      "chart-cell",
       {
-        data: values,
-        backgroundColor: barBackgroundColors,
-        barPercentage: 1.09,
+        "chart-cell-highlight": (highlightedItemKeys || []).includes(
+          entry[keyField]
+        ),
+        "chart-cell-active": highlightActive && dataIndex === activeItemIndex,
       },
-    ];
-
-    const dataForChart = { datasets: datasets, labels: labels };
-
-    const options: ChartOptions<"bar"> = {
-      maintainAspectRatio: false,
-      indexAxis: "y",
-      responsive: true,
-      scales: {
-        x: {
-          border: {
-            color: "black",
-          },
-          grid: {
-            display: true,
-            drawOnChartArea: false,
-            drawTicks: true,
-            tickLength: 7,
-            color: "black",
-          },
-          title: {
-            display: true,
-            text: xLabel,
-            color: "black",
-          },
-          ticks: {
-            color: "black",
-          },
-        },
-        y: {
-          border: {
-            color: "black",
-          },
-          grid: {
-            offset: false,
-            display: true,
-            drawOnChartArea: false,
-            drawTicks: true,
-            tickLength: 7,
-            color: "black",
-          },
-          ticks: {
-            color: "#1D70B8",
-            font: (context) => {
-              const label = context.tick.label;
-              const weight =
-                labels[chosenSchoolIndex] === label ? "bolder" : "normal";
-              return {
-                weight: weight,
-              };
-            },
-          },
-        },
-      },
-    };
-
-    useImperativeHandle(ref, () => ({
-      download() {
-        if (chartRef.current) {
-          const a = document.createElement("a");
-          a.href = chartRef.current.toBase64Image();
-          a.download = `${chartName}.png`;
-          a.click();
-        }
-      },
-    }));
+      `chart-cell-series-${seriesIndex}`,
+      config?.className
+    );
 
     return (
-      <div style={chartContainerStyle}>
-        <Bar
-          aria-label={`Bar chart showing ${chartName} as ${xLabel}`}
-          data={dataForChart}
-          options={options}
-          plugins={[underLinePlugin]}
-          ref={chartRef}
-        />
-      </div>
+      <Cell
+        key={`cell-${entry[keyField]}`}
+        cursor={highlightActive ? "pointer" : "default"}
+        className={className}
+      />
     );
-  }
-);
+  };
+
+  return (
+    // a11y: https://github.com/recharts/recharts/issues/3816
+    <div
+      aria-label={chartName}
+      className="govuk-body-s govuk-!-font-size-14 full-height-width"
+      role="img"
+    >
+      <ResponsiveContainer>
+        <BarChart
+          barCategoryGap={barCategoryGap}
+          data={data}
+          layout="vertical"
+          margin={{
+            top: margin,
+            right: margin + (labels ? 20 : 0),
+            bottom: margin,
+            left: margin,
+          }}
+          onMouseMove={handleBarChartMouseMove}
+          ref={rechartsRef}
+        >
+          {grid && <CartesianGrid />}
+          {!!tooltip && <Tooltip content={tooltip} />}
+          {visibleSeriesNames.map((seriesName, seriesIndex) => (
+            <Bar key={seriesName as string} dataKey={seriesName as string}>
+              {data.map((entry, dataIndex) =>
+                renderCell(entry, dataIndex, seriesName, seriesIndex)
+              )}
+              {labels && (
+                <LabelList
+                  dataKey={seriesName as string}
+                  content={(c) => (
+                    <LabelListContent
+                      {...c}
+                      valueFormatter={
+                        seriesConfig?.[seriesName]?.valueFormatter
+                      }
+                    />
+                  )}
+                />
+              )}
+            </Bar>
+          ))}
+          <XAxis
+            type="number"
+            hide={hideXAxis}
+            tickFormatter={(value) =>
+              valueFormatter
+                ? valueFormatter(value, { valueUnit })
+                : String(value)
+            }
+          >
+            {valueLabel && (
+              <Label value={valueLabel} offset={0} position="bottom" />
+            )}
+          </XAxis>
+          <YAxis
+            dataKey={seriesLabelField as string}
+            hide={hideYAxis}
+            interval={0}
+            tick={tick}
+            type="category"
+            width={tickWidth}
+          ></YAxis>
+          {legend && (
+            <Legend
+              align="right"
+              verticalAlign="top"
+              formatter={(value) =>
+                (seriesConfig && seriesConfig[value]?.label) || value
+              }
+              height={30}
+            />
+          )}
+        </BarChart>
+      </ResponsiveContainer>
+    </div>
+  );
+}
+
+function LabelListContent(props: LabelListContentProps) {
+  const { valueFormatter, height, value, width, x, y } = props;
+
+  return (
+    <g>
+      <Text
+        x={(x as number) + (width as number) + (height as number) / 2}
+        y={(y as number) + (height as number) / 2}
+        textAnchor="start"
+        dominantBaseline="middle"
+      >
+        {valueFormatter
+          ? valueFormatter(value as ChartSeriesValue)
+          : String(value)}
+      </Text>
+    </g>
+  );
+}
+
+// https://stackoverflow.com/a/58473012/504477
+export const HorizontalBarChart = forwardRef(HorizontalBarChartInner) as <
+  TData extends ChartDataSeries,
+>(
+  p: HorizontalBarChartProps<TData> & { ref?: Ref<ChartHandler> }
+) => ReactElement;
